finish migrating basic stuff of custom vms to local
