+++ server/network.go
@@ -293,7 +293,7 @@ func (lc *localNetwork) start() error {
 
 func (lc *localNetwork) startWait(
 	argCtx context.Context,
-	chainSpecs []local.BlockchainSpec, // VM name + genesis bytes
+	chainSpecs []network.BlockchainSpec, // VM name + genesis bytes
 	readyCh chan struct{}, // messaged when initial network is healthy, closed when subnet installations are complete
 ) {
 	defer close(lc.startDoneCh)
@@ -316,7 +316,7 @@ func (lc *localNetwork) startWait(
 
 func (lc *localNetwork) createBlockchains(
 	argCtx context.Context,
-	chainSpecs []local.BlockchainSpec, // VM name + genesis bytes
+	chainSpecs []network.BlockchainSpec, // VM name + genesis bytes
 	createBlockchainsReadyCh chan struct{}, // closed when subnet installations are complete
 ) {
 	// createBlockchains triggers a series of different time consuming actions
@@ -335,13 +335,7 @@ func (lc *localNetwork) createBlockchains(
 		return
 	}
 
-	chainInfos, err := lc.installCustomVMs(ctx, chainSpecs)
-	if err != nil {
-		lc.startErrCh <- err
-		return
-	}
-
-	if err := lc.waitForCustomVMsReady(ctx, chainInfos); err != nil {
+	if err := lc.nw.CreateBlockchains(ctx, chainSpecs); err != nil {
 		lc.startErrCh <- err
 		return
 	}
@@ -375,8 +369,7 @@ func (lc *localNetwork) createSubnets(
 		return
 	}
 
-	_, err := lc.setupWalletAndInstallSubnets(ctx, numSubnets)
-	if err != nil {
+	if err := lc.nw.CreateSubnets(ctx, numSubnets); err != nil {
 		lc.startErrCh <- err
 		return
 	}
